Improve ADC with timer2 usage, modify rounding method of timer reload value
The analog to digital conversion must be driven by a timer
module. The timer module's overflow impulse starts the
ADC module's conversion. The timer overflow frequency
can not be higher than the reciprocal of the conversion
time (with the post tracking time).

The ADC application strategy could be modified; when the
R resistance is too big (1M ohm) then the current solving
strategy fails. The lowest system clock could be divided
to a lower value to increase post tracking time. Or the
system clock could be kept in a high value (24.5 MHz) and
the AD0SC register value could be set to a higher value,
since there is no restriction in C8051F410 data sheet for
minimum sar clock, e.g.: 24.5MHz / 512