Support external system clock usage
The backend and frontend needed an update to handle the
case when the system clock is provided by an external
oscillator. This affects timer (C code generation for
external oscillator is not supported yet) and uart
applications (internal oscillator have less accuracy and
a message have to be shown to the user).